Print actual Debian package filename
The current package_debian.sh circumvents to print the
full filename of the Debian package by using wildcards.

However, this is not very user-friendly and can lead to
confusion or bugs. Especially when you do a remote or
Docker build, the filename should be explictly printed,
otherwise it can be difficult to find the result.

Therefore, assemble the full filename.

Signed-off-by: Florian Kauer <florian.kauer@linutronix.de>